tasks/object-detection/vlm-approach/pyproject.toml
[project]
name = "vlm-approach"
version = "0.1.0"
description = "VLM-based classification pipeline for object detection challenge"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "timm>=1.0.0",
    "Pillow>=10.0.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.5.0",
    "ultralytics>=8.1.0",
    "unsloth[cu124]>=2026.3.8",
    "wandb>=0.25.1",
    "onnxruntime-gpu>=1.24.4",
    "datasets>=3.0.0",
]
